requirements.txt
mysql-connector-python
django
flake8